% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{at.least.n}
\alias{at.least.n}
\title{at.least.n}
\usage{
at.least.n(vector, n)
}
\arguments{
\item{vector}{numeric vector to test.}

\item{n}{number of elements in vector that must be > 0 for the function to return TRUE.}
}
\description{
Determines if at least n values in a vector are higher than 0.
}
\examples{
keeps the rows in exprs that have at least 6 values > 0 (6 samples are > 0)
exprs <- exprs[apply(exprs[,1:nrow(s2c)], 1, function(x) at.least.n(x,6)),]
}
\keyword{filter}
