% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{at_least_counts_in_samples}
\alias{at_least_counts_in_samples}
\title{at_least_counts_in_samples}
\usage{
at_least_counts_in_samples(vector, c, s)
}
\arguments{
\item{vector}{numeric vector to test.}

\item{c}{minimum number of counts there must be in a sample for the function to return TRUE.}

\item{s}{minimum number of samples that must have "c" number of counts for the function to return TRUE.}
}
\description{
Determines if at least "s" samples have at least "c" counts in a vector.
}
\examples{
keeps the rows in exprs that have at least 5 counts in 4 samples
exprs <- exprs[apply(exprs[,1:nrow(s2c)], 1, function(x) at_least_counts_in_samples(x, 5, 4)),]
}
\keyword{filter}
