% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{gg.cols}
\alias{gg.cols}
\title{gg.cols}
\usage{
gg.cols(n)
}
\arguments{
\item{n}{number of colors to return.}
}
\description{
Return a ggplot palette with n colors
}
\examples{
cols = gg.cols(6)
}
\keyword{palette}
