% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{gg_cols}
\alias{gg_cols}
\title{gg_cols}
\usage{
gg_cols(n)
}
\arguments{
\item{n}{number of colors to return.}
}
\description{
Return a ggplot palette with n colors
}
\examples{
cols = gg_cols(6)
}
\keyword{palette}
