% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{mart.data.import}
\alias{mart.data.import}
\title{mart.data.import}
\usage{
mart.data.import()
}
\description{
Downloads or loads the biomart annotated reference of genes and transcripts (t2g)
and the compatec gene to gene symbol reference (t2g.annot)
}
\examples{
annot <- mart.data.import()
t2g <- annot$t2g
t2g.annot <- annot$t2g.annot
}
\keyword{mart}
