% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{p2g.import}
\alias{p2g.import}
\title{p2g.import}
\usage{
p2g.import()
}
\description{
Downloads or loads the biomart annotated reference of protein to gene (p2g)
}
\examples{
p2g <- p2g.import()
}
\keyword{mart}
