% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{std.err}
\alias{std.err}
\title{std.err}
\usage{
std.err(x)
}
\arguments{
\item{x}{numeric vector to test.}
}
\description{
Calculate the standard error of a vector x
}
\examples{
sderr = std.err(x)
}
\keyword{sd}
