% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArpyLib.R
\name{go_terms}
\alias{go_terms}
\title{go_terms}
\usage{
go_terms(gene_list, universe)
}
\arguments{
\item{universe}{background list to test against.}

\item{gene.list}{list of genes to be tested.}
}
\description{
Functional annotation of a list of genes (GOprofiler)
}
\examples{
gene.GO <- go_terms(gene.list=limma.deg, universe=rownames(norm.counts))
}
\keyword{annotation}
